Refactor infer deps to separate namespace
